:file_folder: refactor: Detach type methods in 2-files separated by flag
Has a problem with generify duck-interface not found any method
